+++ lib/main.dart
@@ -72,6 +72,22 @@ void main() async {
     assetsIconPath: 'assets/cover.png',
   );
 
+  if (TPlatform.isDesktop) {
+    WindowOptions windowOptions = const WindowOptions(
+      size: Size(514, 414), // စတင်ဖွင့်တဲ့အချိန် window size
+
+      backgroundColor: Colors.transparent,
+      skipTaskbar: false,
+      center: false,
+      title: "Novel V3",
+    );
+
+    windowManager.waitUntilReadyToShow(windowOptions, () async {
+      await windowManager.show();
+      await windowManager.focus();
+    });
+  }
+
   runApp(
     MultiProvider(
       providers: [